Refactoring: Ordering the source code by meaning (part 3) (#24756)
