+++ ydb/core/persqueue/common/sourceid_info.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
+#include <ydb/core/persqueue/public/partition_key_range/partition_key_range.h>
 #include "heartbeat.h"
 #include <ydb/core/protos/pqconfig.pb.h>
