--- ydb/core/persqueue/partition_key_range/ut/ya.make
@@ -1,5 +1,0 @@
-UNITTEST_FOR(ydb/core/persqueue/partition_key_range)
-
-SRCS(partition_key_range_sequence_ut.cpp)
-
-END()
