+++ ydb/core/persqueue/pqrb/partition_scale_manager.cpp
@@ -2,7 +2,7 @@
 #include "read_balancer_log.h"
 #include "partition_scale_manager_graph_cmp.h"
 
-#include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
+#include <ydb/core/persqueue/public/partition_key_range/partition_key_range.h>
 #include <fmt/format.h>
 #include <algorithm>
 #include <ranges>
