+++ ydb/core/persqueue/pqrb/ya.make
@@ -18,7 +18,6 @@ PEERDIR(
     ydb/core/engine/minikql
     ydb/core/persqueue/events
     ydb/core/persqueue/common
-    ydb/core/persqueue/partition_key_range
 )
 
 END()
