+++ ydb/core/persqueue/pqtablet/partition/sourceid.h
@@ -4,7 +4,7 @@
 #include <ydb/core/persqueue/common/sourceid_info.h>
 #include <ydb/core/persqueue/common/key.h>
 #include "ownerinfo.h"
-#include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
+#include <ydb/core/persqueue/public/partition_key_range/partition_key_range.h>
 #include <ydb/core/protos/pqconfig.pb.h>
 
 #include <util/generic/set.h>
