+++ ydb/core/persqueue/public/partition_index_generator/partition_index_generator_ut.cpp
@@ -1,4 +1,4 @@
-#include <ydb/core/persqueue/partition_index_generator/partition_index_generator.h>
+#include "partition_index_generator.h"
 #include <library/cpp/testing/unittest/registar.h>
 
 namespace NKikimr::NPQ {
