+++ ydb/core/persqueue/public/partition_index_generator/ya.make
