+++ ydb/core/persqueue/public/partition_key_range/ut/ya.make
@@ -1,0 +1,5 @@
+UNITTEST_FOR(ydb/core/persqueue/public/partition_key_range)
+
+SRCS(partition_key_range_sequence_ut.cpp)
+
+END()
