+++ ydb/core/persqueue/public/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    config.cpp
     pq_database.cpp
     pq_rl_helpers.cpp
     utils.cpp
@@ -20,5 +21,7 @@ RECURSE(
     counters
     fetcher
     list_topics
+    partition_index_generator
+    partition_key_range
     write_meta
 )
