+++ ydb/core/persqueue/writer/partition_chooser_impl.cpp
@@ -1,7 +1,7 @@
 #include "partition_chooser_impl.h"
 
 #include <library/cpp/digest/md5/md5.h>
-#include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
+#include <ydb/core/persqueue/public/partition_key_range/partition_key_range.h>
 #include <ydb/core/persqueue/public/utils.h>
 #include <ydb/services/lib/sharding/sharding.h>
