+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_kesus.cpp
@@ -4,7 +4,7 @@
 
 #include <ydb/core/base/subdomain.h>
 #include <ydb/core/kesus/tablet/events.h>
-#include <ydb/core/persqueue/config/config.h>
+#include <ydb/core/persqueue/public/config.h>
 
 namespace {
