+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_pq.cpp
@@ -5,10 +5,10 @@
 
 #include <ydb/core/base/subdomain.h>
 #include <ydb/core/mind/hive/hive.h>
-#include <ydb/core/persqueue/config/config.h>
-#include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
-#include <ydb/core/persqueue/partition_key_range/partition_key_range_sequence.h>
-#include <ydb/core/persqueue/partition_index_generator/partition_index_generator.h>
+#include <ydb/core/persqueue/public/config.h>
+#include <ydb/core/persqueue/public/partition_key_range/partition_key_range.h>
+#include <ydb/core/persqueue/public/partition_key_range/partition_key_range_sequence.h>
+#include <ydb/core/persqueue/public/partition_index_generator/partition_index_generator.h>
 #include <ydb/core/persqueue/public/utils.h>
 
 #include <ydb/services/lib/sharding/sharding.h>
