+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_subdomain.cpp
@@ -4,7 +4,7 @@
 #include "schemeshard_impl.h"
 
 #include <ydb/core/base/subdomain.h>
-#include <ydb/core/persqueue/config/config.h>
+#include <ydb/core/persqueue/public/config.h>
 
 namespace {
