+++ ydb/core/tx/schemeshard/schemeshard__operation_create_bsv.cpp
@@ -5,7 +5,7 @@
 
 #include <ydb/core/base/subdomain.h>
 #include <ydb/core/mind/hive/hive.h>
-#include <ydb/core/persqueue/config/config.h>
+#include <ydb/core/persqueue/public/config.h>
 
 namespace {
