+++ ydb/core/tx/ya.make
@@ -12,7 +12,6 @@ SRCS(
 PEERDIR(
     ydb/library/actors/core
     ydb/core/base
-    ydb/core/persqueue/config
     ydb/core/protos
     ydb/core/tablet
     ydb/core/tablet_flat
