+++ ydb/services/datastreams/ya.make
@@ -16,6 +16,7 @@ PEERDIR(
     ydb/core/grpc_services
     ydb/core/mind
     ydb/core/persqueue/public/list_topics
+    ydb/core/persqueue/public/partition_key_range
     ydb/core/persqueue/public/write_meta
     ydb/public/api/grpc
     ydb/public/api/grpc/draft
